DDT-634 moved cancel methods to BKGController
