+++ src/main/java/org/dcsa/bkg/controller/BKGConfirmedBookingsController.java
@@ -3,10 +3,12 @@ package org.dcsa.bkg.controller;
 import lombok.RequiredArgsConstructor;
 import org.dcsa.bkg.model.transferobjects.BookingConfirmationTO;
 import org.dcsa.bkg.service.BookingService;
-import org.springframework.http.HttpStatus;
 import org.springframework.http.MediaType;
 import org.springframework.validation.annotation.Validated;
-import org.springframework.web.bind.annotation.*;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.PathVariable;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RestController;
 import reactor.core.publisher.Mono;
 
 import javax.validation.constraints.Size;
@@ -28,19 +30,4 @@ public class BKGConfirmedBookingsController {
     return bookingService.getBookingConfirmationByCarrierBookingReference(carrierBookingReference);
   }
 
-  // To avoid spelling confusion we just accept both spellings
-  @PostMapping(path = "{carrierBookingReference}/cancelation")
-  @ResponseStatus(HttpStatus.NO_CONTENT)
-  public Mono<Void> bookingCancelation(
-      @PathVariable @Size(max = 35) String carrierBookingReference) {
-    // ToDo: adjust this when the IM is ready for booking
-    return bookingCancellation(carrierBookingReference);
-  }
-
-  @PostMapping(path = "{carrierBookingRequestReference}/cancellation")
-  @ResponseStatus(HttpStatus.NO_CONTENT)
-  public Mono<Void> bookingCancellation(
-      @PathVariable @Size(max = 35) String carrierBookingRequestReference) {
-    return bookingService.cancelBookingByCarrierBookingReference(carrierBookingRequestReference);
-  }
 }
