+++ src/test/java/org/dcsa/bkg/controller/BKGConfirmedBookingsControllerTest.java
@@ -8,7 +8,6 @@ import org.dcsa.core.events.model.enums.DCSATransportType;
 import org.dcsa.core.events.model.enums.LocationType;
 import org.dcsa.core.events.model.enums.PaymentTerm;
 import org.dcsa.core.events.model.transferobjects.LocationTO;
-import org.dcsa.core.exception.UpdateException;
 import org.dcsa.core.exception.handler.GlobalExceptionHandler;
 import org.dcsa.core.security.SecurityConfig;
 import org.junit.jupiter.api.BeforeEach;
@@ -141,25 +140,6 @@ class BKGConfirmedBookingsControllerTest {
   }
 
   @Test
-  @DisplayName(
-      "Get confirmed bookings should return valid list of booking request summaries for valid request.")
-  void confirmedBookingsCancellationShouldReturn204() {
-
-    WebTestClient.ResponseSpec exchange =
-        webTestClient
-            .post()
-            .uri(
-                CONFIRMED_BOOKING_ENDPOINT
-                    + "/"
-                    + bookingConfirmationTO.getCarrierBookingReference()
-                    + "/cancelation")
-            .accept(MediaType.APPLICATION_JSON)
-            .exchange();
-
-    checkStatus204.apply(exchange);
-  }
-
-  @Test
   @DisplayName("POST booking should return 400 for invalid request.")
   void postBookingsShouldReturn400ForInValidBookingRequest() {
 
@@ -169,49 +149,6 @@ class BKGConfirmedBookingsControllerTest {
             .uri(CONFIRMED_BOOKING_ENDPOINT + "/" + "y".repeat(36))
             .accept(MediaType.APPLICATION_JSON)
             .exchange();
-
-    checkStatus400.apply(exchange);
-  }
-
-  @Test
-  @DisplayName(
-      "Get confirmed bookings should return valid list of booking request summaries for valid request.")
-  void confirmedBookingsCancelationShouldReturn204() {
-
-    WebTestClient.ResponseSpec exchange =
-        webTestClient
-            .post()
-            .uri(
-                CONFIRMED_BOOKING_ENDPOINT
-                    + "/"
-                    + bookingConfirmationTO.getCarrierBookingReference()
-                    + "/cancellation")
-            .accept(MediaType.APPLICATION_JSON)
-            .exchange();
-    checkStatus204.apply(exchange);
-  }
-
-  @Test
-  @DisplayName("Cancelling of a booking in unallowed status should return a 400 for invalid request")
-  void confirmedBookingsCancellationShouldReturn400() {
-
-    Mockito.when(
-        bookingService.cancelBookingByCarrierBookingReference(
-          bookingConfirmationTO.getCarrierBookingReference()))
-      .thenReturn(Mono.error(
-        new UpdateException(
-          "Cannot Cancel Booking that is not in status RECE, PENU or CONF")));
-
-    WebTestClient.ResponseSpec exchange =
-      webTestClient
-        .post()
-        .uri(
-          CONFIRMED_BOOKING_ENDPOINT
-            + "/"
-            + bookingConfirmationTO.getCarrierBookingReference()
-            + "/cancellation")
-        .accept(MediaType.APPLICATION_JSON)
-        .exchange();
 
     checkStatus400.apply(exchange);
   }
